+++ spec/integration/braintree/oauth_gateway_spec.js
@@ -122,6 +122,7 @@ describe('OAuthGateway', function () {
         scope: 'read_write',
         state: 'baz_state',
         landingPage: 'login',
+        loginOnly: 'true',
         user: {
           country: 'USA',
           email: 'foo@example.com',
@@ -159,7 +160,7 @@ describe('OAuthGateway', function () {
 
       let query = function (searchKey) { // eslint-disable-line func-style
         let parts = queryString.split('&');
-        let foundValue = null;
+        let foundValue;
 
         parts.forEach(function (part) {
           let parts = Array.from(part.split('='));
@@ -192,6 +193,7 @@ describe('OAuthGateway', function () {
       assert.equal(query('scope'), 'read_write');
       assert.equal(query('state'), 'baz_state');
       assert.equal(query('landing_page'), 'login');
+      assert.equal(query('login_only'), 'true');
 
       assert.equal(query('user[country]'), 'USA');
       assert.equal(query('business[name]'), '14 Ladders');
@@ -240,7 +242,7 @@ describe('OAuthGateway', function () {
 
       let query = function (searchKey) { // eslint-disable-line func-style
         let parts = queryString.split('&');
-        let foundValue = null;
+        let foundValue;
 
         parts.forEach(function (part) {
           let parts = Array.from(part.split('='));