+++ spec/integration/braintree/payment_method_gateway_spec.js
@@ -9,7 +9,7 @@ let ValidationErrorCodes = require('../../../lib/braintree/validation_error_code
 
 describe('PaymentMethodGateway', function () {
   describe('create', function () {
-    let customerId = null;
+    let customerId;
 
     it('works with an unknown payment method nonce', done =>
       specHelper.defaultGateway.customer.create({firstName: 'John', lastName: 'Smith'}, function (err, response) {
@@ -911,7 +911,18 @@ describe('PaymentMethodGateway', function () {
                   orderId: 'merchant-order-id',
                   customField: 'custom merchant field',
                   description: 'merchant description',
-                  amount: '1.23'
+                  amount: '1.23',
+                  shipping: {
+                    company: 'Braintree',
+                    countryName: 'United States of America',
+                    extendedAddress: 'Apt B',
+                    firstName: 'first',
+                    lastName: 'last',
+                    locality: 'Chicago',
+                    postalCode: '60646',
+                    region: 'IL',
+                    streetAddress: '123 Fake St'
+                  }
                 }
               }
             };
@@ -1128,7 +1139,7 @@ describe('PaymentMethodGateway', function () {
 
   describe('find', function () {
     context('credit card', function () {
-      let paymentMethodToken = null;
+      let paymentMethodToken;
 
       before(done =>
         specHelper.defaultGateway.customer.create({firstName: 'John', lastName: 'Smith'}, function (err, response) {
@@ -1949,7 +1960,7 @@ describe('PaymentMethodGateway', function () {
   });
 
   describe('delete', function () {
-    let paymentMethodToken = null;
+    let paymentMethodToken;
 
     context('credit card', function () {
       before(done =>
@@ -2053,8 +2064,7 @@ describe('PaymentMethodGateway', function () {
   });
 
   context('grant and revoke payment methods', function () {
-    let creditCard = null;
-    let grantingGateway = null;
+    let creditCard, grantingGateway;
 
     before(function (done) {
       let partnerMerchantGateway = braintree.connect({